* switch RLE chunks to use 8 bits for the runs, since very wide sprites are not   as common (saves memory, but wide sprites render slower)

+++ src/engine/render/rle.h
@@ -5,12 +5,12 @@
 
 struct RLEChunk {
     byte col;
-    uint16 length;
+    byte length;
 };
 
 struct MonoRLEChunk {
-    uint16 bglen;
-    uint16 fglen;  
+    byte bglen;
+    byte fglen;  
 };
 
 typedef struct RLEChunk     far RLEImage;
